+++ Algos/Sparse_Table.cpp
@@ -1,4 +1,5 @@
 #include <bits/stdc++.h>
+
 using namespace std;
 
 const int MAX_N = 100'005;
@@ -7,13 +8,13 @@ vector<int> a(MAX_N, 0);
 vector<vector<int>> m(MAX_N, vector<int>(LOG, 0));
 vector<int> bin_log;
 
-int queryMin(int L, int R) // O(1)
-{
+int queryMin(int L, int R) {
     int i = bin_log[R - L + 1];
     int minimum = min(m[i][L], m[i][R - (1 << i) + 1]);
     return minimum;
 }
-int querySum(int L, int R) {
+
+long long querySum(int L, int R) {
     long long sum = 0;
     for (int i = LOG; i >= 0; i--) {
 	if ((1 << i) <= R - L + 1) {
@@ -36,14 +37,12 @@ int main() {
 	cin >> a[i];
 	m[0][i] = a[i];
     }
-    // 2) preprocessing for min, O(N*log(N))
-    // for (int i = 1; i <= LOG; i++){
-    // 		for (int j = 0; j + (1 << i) <= n; j++){
-    // 			m[i][j] = min(m[i - 1][j], m[i - 1][j + (1 << (i -
-    // 1))]);
-    // 		}
-    // }
-
+    /*    2) preprocessing for min, O(N*log(N))*/
+    /*    for (int i = 1; i <= LOG; i++){*/
+    /*for (int j = 0; j + (1 << i) <= n; j++) {*/
+    /*    m[i][j] = min(m[i - 1][j], m[i - 1][j + (1 << (i - 1))]);*/
+    /*}*/
+    /*   }*/
     for (int i = 1; i <= LOG; i++) {
 	for (int j = 0; j + (1 << i) <= n; j++) {
 	    m[i][j] = m[i - 1][j] + m[i - 1][j + (1 << (i - 1))];
@@ -55,22 +54,19 @@ int main() {
     for (int i = 0; i < q; i++) {
 	int L, R;
 	cin >> L >> R;
-	// cout << queryMin(L, R) << "\n";
+	/*cout << queryMin(L, R) << "\n";*/
 	cout << querySum(L, R) << "\n";
     }
-    // for (auto &&a : m)
-    // {
-    // 	for (auto &&b : a)
-    // 	{
-    // 		if (b!=0)
-    // 		{
-    // 			cout << b << " ";
-    // 		}
-    // 	}
-    // 	cout << "\n";
-    // }
-    // for(auto&& a : bin_log){
-    // 	cout<<a<<" ";
-    // }
+    /*   for (auto &&a : m) {*/
+    /*for (auto &&b : a) {*/
+    /*    if (b != 0) {*/
+    /*	cout << b << " ";*/
+    /*    }*/
+    /*}*/
+    /*cout << "\n";*/
+    /*   }*/
+    /*   for (auto &&a : bin_log) {*/
+    /*cout << a << " ";*/
+    /*   }*/
     return 0;
 }
